add index file and fix some thing
